fix prefer-const issues, split out makePermutations
