+++ src/circuit/.eslintrc.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: ['../../.eslintrc.cjs'],
+  rules: {
+    'no-autofix/prefer-const': 'off',
+  },
+};
