+++ src/circuit/range.ts
@@ -1,0 +1,5 @@
+export default function* range(start: number, end: number) {
+  for (let i = start; i < end; i++) {
+    yield i;
+  }
+}
